made maps tab prettier and added controls restrictions to map

+++ app/src/main/java/edu/ucsb/cs/cs184/trungbui/cs184project/ui/google_maps/Question.kt
@@ -2,8 +2,8 @@ package edu.ucsb.cs.cs184.trungbui.cs184project.ui.google_maps
 
 data class Question(
     val id: Int,
-    val lat: Float,
-    val long: Float,
+    val lat: Double,
+    val long: Double,
     val optionOne: String,
     val optionTwo: String,
     val optionThree: String,
